Change roll api, add config file
Changes the Mythal::Roll api to return a custom response object. This
will enable reuse of the roll service, and the output for the CLI now
resides in DiceRollResult#message.

Adds a default_data.yml and configuration class for the gem. Data will
no longer be hard coded in the lib files, but loaded from external
sources.

bump version